Related to #82. Make track posting more robust

+++ sequencer/localpersistence/src/main/java/sequencer/persistence/TrackMetaData.java
@@ -3,5 +3,5 @@ package sequencer.persistence;
 /**
  * Record for metadata of track object. Mainly used for search results
  */
-public record TrackMetaData(String id, String name, String artist, int timestamp) {
+public record TrackMetaData(String id, String name, String artist, long timestamp) {
 }
